+++ app/certificate/views.py
@@ -1,0 +1,82 @@
+from flask import url_for, redirect
+from flask_login import login_required
+
+from . import certificate
+from .certificate import generate_certificate, 
+
+CERT_ROUTES = {
+    "cancer-pulmon": "plmn",
+    "cancer-mama": "mama",
+    "cancer-cervicouterino": "crvu",
+    "cancer-prostata": "psta",
+    "cancer-testicular": "tstc"
+}
+
+
+@certificate.route("/<name>")
+@login_required
+def download_certificate(name):
+
+    # Validate that a valid url argument for course name was entered
+    if name not in CERT_ROUTES:
+        return redirect(url_for("main.index"))
+
+    # Shorthand for current_user
+    cu = current_user
+
+    # Fetch the current user from the database
+    user = User.objects(email=cu.email).first()
+    if user == None:
+        return redirect(url_for("main.index"))
+
+    # Check if user has completed the quiz
+    if not user.has_passed_quiz(CERT_ROUTES[name]):
+        flash("Debe completar la evaluación al final del módulo para obtener su certificado.")
+        return redirect(url_for("cursos.{}".format(name.replace('-', '_'))))
+
+    
+    
+    
+    
+    # Current working dir to set abs path for pdfkit html resources
+    cwd = os.getcwd()
+
+    # Get the certificate data: title, user name, date, course, background image and font path
+    cert_title = "Certificado Campus Virtual - {} {} {}".format(
+        cu.first_name, cu.paternal_last_name, cu.maternal_last_name)
+    cert_name = (
+        cu.first_name + ' ' + cu.paternal_last_name + ' ' + cu.maternal_last_name).upper()
+    cert_date = datetime.date.today()
+    cert_date = "{}/{}/{}".format(cert_date.day,
+                                  cert_date.month, cert_date.year)
+    cert_course_name = COURSE_CERT[name]["cert_course_name"]
+    cert_bg_img = os.path.join(
+        cwd, "app/static/img", COURSE_CERT[name]["cert_bg_img"]).replace('\\', '/')
+    cert_font_path = os.path.join(
+        cwd, "app/static/css/fonts/").replace('\\', '/')
+
+    # The Jinja rendered string to pass to pdf generator
+    rendered = render_template("certificate/_certificate.html",
+                               title=cert_title,
+                               name=cert_name,
+                               date=cert_date,
+                               course_name=cert_course_name,
+                               background=cert_bg_img,
+                               font_path=cert_font_path)
+
+    # Generate pdf payload using pdfkit
+    pdf = pdfkit.from_string(
+        input=rendered, output_path=False, configuration=pdfkit_config,
+        options={
+            "enable-local-file-access": None,
+            "disable-smart-shrinking": None
+        }
+    )
+
+    # Set up the pdf response headers for a pdf file instead of regular html
+    response = make_response(pdf)
+    response.headers["Content-Type"] = "application/pdf"
+    response.headers["Content-Disposition"] = "inline; filename={}.pdf". format(
+        cert_title)
+
+    return response
